towards-targets: Move OmpizerGPU to gpu.py
