+++ devito/targets/common/__init__.py
@@ -1,5 +1,5 @@
 from .engine import *  # noqa
 from .blocking import *  # noqa
-from .openmp import NThreads, NThreadsNested, NThreadsNonaffine, Ompizer, OmpizerGPU  # noqa
+from .openmp import *  # noqa
 from .rewriters import *  # noqa
 from .definitions import *  # noqa
